Fixed Cannon ball not damaging on first hit.
The cannon ball was not damaging the player when it hit the player for the first time. Fixed the bug by setting "remainingHitTime" variable to "hitTime" variable in the start method, so that when the player is hit for the first time he takes damage and the code does not directly skip to resetting the wasHit variable.